Split out carbon emissions from cap (#674)
Splitting out carbon emissions from the cap allows for outputting
carbon emissions for all fuel projects, regardless of whether you
are running a carbon cap or not.

This required adding a carbon_emissions_rule to each operational
type. In the future, this rule could be used to provide different
and/or simplified carbon emission calculations that could possibly
avoid the use of fuels entirely.