Move dynamic component list alterations
Instead of doing them in add_model_components, build the lists in
determine_dynamic_components so that you can access these when
you process results.

We also move tx capacity cost aggregations to the objective
folder to be consistent with the other aggregations.

This is a pre-requisite update for outputting total system
costs by component.