Refactor operational types (#692)
This is a refactor of the operational types to streamline data flows 
and remove redundant code. There are two main ideas:

1. All cost and fuel-related parameters are moved to where they are 
actually used, in the respective modules under 
gridpath.project.operations that call methods from the operational 
types. We now create the cost and fuel burn components and outputs 
only for projects for which the respective parameters are specified.
2. The gridpath.project.operations.operational_types.__init__ file now 
contains defaults for each operational type method. If an op-type 
module does not include a particular method, the default is used, 
allowing us to remove a lot of the repetitive code in the op type 
modules.

A number of tests had to be moved around and I added a little bit of 
testing functionality.

A few objective functions changed very slightly. I looked into it and 
it was a slightly different decision in how CCGT and CCGT_New are 
dispatched, resulting in the same total energy, but a small change in 
fuel use due to the heat rate curves for the two. I'm pretty sure this 
is just the solver stopping at a very slightly different solution 
after the change in the number of components added.