Support deleting prior data on CSV-to-DB import (#698)
Support for deleting prior data for an individual subscenario ID or 
project-subscenario ID before re-import of data.

If any scenarios use the data, confirm with the user they want to 
delete and reimport to avoid mismatch between scenario inputs and 
results.

This required avioding FOREIGN KEY errors by deleting the data in the 
right order for subscenarios consisting of more than one table and 
updating the subscenario values to NULL in the scenarios table if any 
scenarios use the inputs (and in the base table for project-level 
inputs, e.g. the opchar table for variable profiles).