Ensure unit can't be synced when unavailable (#777)
Constrains Synced to be <= Availability_Derate. This is needed to
prevent the model from committing a unit while still unavailable, thus
avoiding startup costs, which are based on available capacity started
up. Note that when the availability derate is <0, this will force
commitment, power, and startup/shutdown power to 0 for gen_commit_bin,
i.e. there are no partial availabilities. For gen_commit_lin, partial
availability is allowed and the unit can still sync but power provision
will be limited based on the availability derate.

This also ensures that the relationship between Synced and Commit in
synced_constraint_rule() is preserved for the special cases when Pmin
is endogenously set to 0 (rather than just setting Synced to 0), so
that the aforementioned behavior does not happen in those cases.