Support for flexible-duration periods (#792)
Define period duration via a start year and end year parameters, and 
calculate period duration (in years) based on those parameters. We 
then can use the period duration to adjust annualized capacity costs
and incur a fraction (or multiple) of the annual costs based on the 
number of years in a period.

Also implement checks on timepoint weights to ensure the sum of 
timepoint duration X weight over all timepoints in a period sums up to
the number of hours in a year. When multiplied by the period duration 
and discount factor, this weights capacity costs and operational costs 
equally in the objective function.