Created instantaneous penetration optional feature  (#1148)
New feature to model the minimum and maximum penetration of a 
resource set at every timepoint. The requirements for this constrain 
are base and timepoint specific power limit, on general a percentage 
of the load and on ratio based on the capacity and power of other 
projects. Also created unit test for this feature.